+++ src/functions/rotateVector.ts
@@ -4,7 +4,7 @@ import { degreesToRadians } from './degreesToRadians'
 /**
  * Rotate a vector counter-clockwise around the origin.
  */
-export function rotate([x, y]: Vector2, degrees: number): Vector2 {
+export function rotateVector([x, y]: Vector2, degrees: number): Vector2 {
   const radians = degreesToRadians(degrees)
   const cos = Math.cos(radians)
   const sin = Math.sin(radians)